# coordinator/flex/server/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: graphscope@alibaba-inc.com
    name: GraphScope
  description: |-
    This is a specification for GraphScope FLEX HTTP service based on the OpenAPI 3.0 specification. You can find out more details about
    specification at [doc](https://swagger.io/specification/v3/).

    Some useful links:
    - [GraphScope Repository](https://github.com/alibaba/GraphScope)
    - [The Source API definition for GraphScope Interactive](https://github.com/GraphScope/portal/tree/main/httpservice)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GraphScope FLEX HTTP SERVICE API
  version: 1.9.2
...
servers:
- url: /
...
paths:
...
  /api/v1/graph:
    get:
      description: List all graphs
      operationId: list_graphs
      responses:
        "200":
          content:
            application/json:
              example:
              - id: GRAPHID
                name: demo
                creation_time: 2024-01-01 00:00:00
                schema_update_time: 2024-01-02 00:00:00
                data_update_time: 2024-01-03 00:00:00
                stored_procedures:
                - id: STORED-PROCEDURE-ID
                  name: stored_procedure_name
                  type: cypher
                  query: MATCH(a) return COUNT(a);
                  library: /path/to/library
                  params:
                  - name: param1
                    type:
                      string:
                        long_text: null
                  returns:
                  - name: return1
                    type:
                      PrimitiveType: DT_SIGNED_INT64
                  bound_graph: demo
                  runnable: true
                schema:
                  vertex_types:
                  - type_id: 0
                    type_name: person
                    properties:
                    - property_id: 0
                      property_name: id
                      property_type:
                        string:
                          long_text: null
                    - property_id: 1
                      property_name: name
                      property_type:
                        string:
                          long_text: null
                    primary_keys:
                    - id
                  - type_id: 1
                    type_name: software
                    properties:
                    - property_id: 0
                      property_name: id
                      property_type:
                        string:
                          long_text: null
                    - property_id: 1
                      property_name: name
                      property_type:
                        string:
                          long_text: null
                    primary_keys:
                    - id
                  edge_types:
                  - type_id: 0
                    type_name: knows
                    vertex_type_pair_relations:
                    - source_vertex: person
                      destination_vertex: person
                      relation: MANY_TO_MANY
                      x_csr_params:
                        edge_storage_strategy: ONLY_IN
                    properties:
                    - property_id: 0
                      property_name: weight
                      property_type:
                        primitive_type: DT_DOUBLE
                  - type_id: 1
                    type_name: created
                    vertex_type_pair_relations:
                    - source_vertex: person
                      destination_vertex: software
                      relation: MANY_TO_MANY
                      x_csr_params:
                        edge_storage_strategy: ONLY_IN
                    properties:
                    - property_id: 0
                      property_name: weight
                      property_type:
                        primitive_type: DT_DOUBLE
              schema:
                items:
                  $ref: '#/components/schemas/GetGraphResponse'
                type: array
          description: Successfully returned all graphs
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
    post:
      description: Create a new graph
      operationId: create_graph
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGraphRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGraphResponse'
          description: The graph was created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
  /api/v1/graph/pgql:
    post:
      description: Create a new graph by providing pgql
      operationId: create_graph_by_pgql
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGraphByPgqlRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGraphResponse'
          description: The graph was created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
  /api/v1/graph/yaml:
    post:
      description: Create a new graph by providing yaml
      operationId: create_graph_by_yaml
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGraphByYamlRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGraphResponse'
          description: The graph was created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
  /api/v1/graph/{graph_id}:
    delete:
      description: Delete graph by ID
      operationId: delete_graph_by_id
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successfully deleted the graph
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
    get:
      description: Get graph by ID
      operationId: get_graph_by_id
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                id: GRAPHID
                name: demo
                creation_time: 2024-01-01 00:00:00
                schema_update_time: 2024-01-02 00:00:00
                data_update_time: 2024-01-03 00:00:00
                stored_procedures:
                - id: STORED-PROCEDURE-ID
                  name: stored_procedure_name
                  type: cypher
                  query: MATCH(a) return COUNT(a);
                  library: /path/to/library
                  params:
                  - name: param1
                    type:
                      string:
                        long_text: null
                  returns:
                  - name: return1
                    type:
                      PrimitiveType: DT_SIGNED_INT64
                  bound_graph: demo
                  runnable: true
                schema:
                  vertex_types:
                  - type_id: 0
                    type_name: person
                    properties:
                    - property_id: 0
                      property_name: id
                      property_type:
                        string:
                          long_text: null
                    - property_id: 1
                      property_name: name
                      property_type:
                        string:
                          long_text: null
                    primary_keys:
                    - id
                  - type_id: 1
                    type_name: software
                    properties:
                    - property_id: 0
                      property_name: id
                      property_type:
                        string:
                          long_text: null
                    - property_id: 1
                      property_name: name
                      property_type:
                        string:
                          long_text: null
                    primary_keys:
                    - id
                  edge_types:
                  - type_id: 0
                    type_name: knows
                    vertex_type_pair_relations:
                    - source_vertex: person
                      destination_vertex: person
                      relation: MANY_TO_MANY
                      x_csr_params:
                        edge_storage_strategy: ONLY_IN
                    properties:
                    - property_id: 0
                      property_name: weight
                      property_type:
                        primitive_type: DT_DOUBLE
                  - type_id: 1
                    type_name: created
                    vertex_type_pair_relations:
                    - source_vertex: person
                      destination_vertex: software
                      relation: MANY_TO_MANY
                      x_csr_params:
                        edge_storage_strategy: ONLY_IN
                    properties:
                    - property_id: 0
                      property_name: weight
                      property_type:
                        primitive_type: DT_DOUBLE
              schema:
                $ref: '#/components/schemas/GetGraphResponse'
          description: Successfully returned the graph
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
...
  /api/v1/graph/{graph_id}/schema/edge:
    post:
      description: Create a edge type
      operationId: create_edge_type
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEdgeType'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successful created the edge type
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
  /api/v1/graph/{graph_id}/schema/edge/{type_name}:
    delete:
      description: Delete edge type by name
      operationId: delete_edge_type_by_name
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: source_vertex_type
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: destination_vertex_type
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successful deleted the edge type
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
  /api/v1/graph/{graph_id}/schema/vertex:
    post:
      description: Create a vertex type
      operationId: create_vertex_type
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVertexType'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successful created a vertex type
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
  /api/v1/graph/{graph_id}/schema/vertex/{type_name}:
    delete:
      description: Delete vertex type by name
      operationId: delete_vertex_type_by_name
      parameters:
      - explode: false
        in: path
        name: graph_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: Successful deleted the vertex type
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      tags:
      - Graph
      x-openapi-router-controller: flex.server.controllers.graph_controller
...
components:
...
  schemas:
    APIResponse:
      type: string
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          description: Error code
          format: int32
          title: code
          type: integer
        message:
          description: Error message
          title: message
          type: string
      required:
      - code
      - message
      title: Error
...
    CreateGraphByPgqlRequest:
      example:
        schema: schema
        name: name
        description: description
      properties:
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        schema:
          title: schema
          type: string
      title: CreateGraphByPgqlRequest
      x-body-name: create_graph_by_pgql_request
...
    CreateGraphResponse:
      example:
        graph_id: graph_id
      properties:
        graph_id:
          title: graph_id
          type: string
      required:
      - graph_id
      title: CreateGraphResponse
...